feat: コードの整形とリファクタリング
- `create_session.rs`:
  - `execute` メソッドの引数を複数行に分割し、可読性を向上。
  - `compute_revoke_count` の呼び出しを複数行に分割。

- `revoke_session.rs`:
  - `new` メソッドの引数を複数行に分割。

- `mod.rs` と `tonic_service.rs`:
  - モジュールのインポート順序を変更し、可読性を向上。

- `vault_grpc.rs`:
  - `get_secret` と `set_secret` メソッドの引数を単一行に統合し、可読性を向上。

- `health.rs`:
  - `oneshot` メソッドの引数を複数行に分割。

- `vault_handler.rs`:
  - 関数の引数を複数行に分割し、可読性を向上。
  - エラーメッセージの生成を改善。

- `access_log_postgres.rs`:
  - `subject` フィールドの初期化を複数行に分割。

- `cached_secret_store.rs`:
  - 不要な空行を削除。

- `secret_store_postgres.rs`:
  - SQLクエリの呼び出しを複数行に分割。

- `vault_secret_store.rs`:
  - テストの可読性を向上させるために、メソッド呼び出しを複数行に分割。

- `access_log.rs`:
  - `new` メソッドの引数を単一行に統合。

- `cache.rs`:
  - メソッド呼び出しを複数行に分割し、可読性を向上。

- `encryption.rs`:
  - 環境変数の取得を単一行に統合。

- `kafka_producer.rs`:
  - エラーハンドリングを改善。

- `main.rs`:
  - `store` と `metrics` の初期化を複数行に分割。

- `delete_secret.rs` と `get_secret.rs`:
  - ログの生成を複数行に分割し、可読性を向上。

- `list_audit_logs.rs` と `rotate_secret.rs`:
  - メソッドの引数を複数行に分割。

- `set_secret.rs`:
  - `execute` メソッドの戻り値を変更し、出力を構造体に統合。

+++ regions/system/server/rust/saga/src/adapter/grpc/saga_grpc.rs
@@ -54,7 +54,7 @@ pub struct ListSagasRequest {
 #[derive(Debug)]
 pub struct ListSagasResponse {
     pub sagas: Vec<SagaStateProto>,
-    pub total_count: i64,
+    pub total_count: i32,
     pub page: i32,
     pub page_size: i32,
     pub has_next: bool,
@@ -325,7 +325,7 @@ impl SagaGrpcService {
 
         let page = req.page;
         let page_size = req.page_size;
-        let has_next = (page as i64 * page_size as i64) < total_count;
+        let has_next = (page as i64 * page_size as i64) < i64::from(total_count);
 
         let saga_protos = sagas
             .into_iter()
