+++ regions/system/server/rust/saga/src/adapter/handler/saga_handler.rs
@@ -88,7 +88,7 @@ pub struct ListSagasResponse {
 
 #[derive(Debug, Serialize, utoipa::ToSchema)]
 pub struct PaginationResponse {
-    pub total_count: i64,
+    pub total_count: i32,
     pub page: i32,
     pub page_size: i32,
     pub has_next: bool,
@@ -242,7 +242,7 @@ pub async fn list_sagas(
         })
         .collect();
 
-    let has_next = (query.page as i64 * query.page_size as i64) < total;
+    let has_next = (query.page as i64 * query.page_size as i64) < i64::from(total);
 
     Ok(Json(ListSagasResponse {
         sagas: saga_responses,
